requirements.txt
# HTTP requests library
requests>=2.31.0

# Server-Sent Events client
sseclient-py>=1.7.2

# Environment variable management
python-dotenv>=1.0.0

# Type hints support (for Python < 3.9)
typing-extensions>=4.8.0

# Date parsing and handling
python-dateutil>=2.8.2

# URL handling
urllib3>=2.0.7
